+++ scripts/git-versioning.sh
@@ -3,7 +3,7 @@
 version="${1}"
 old_version="${2}"
 
-committer_name=${COMMITTER_NAME:-"IONOS Cloud Docker Machine Driver Robot"}
+committer_name=${COMMITTER_NAME:-"IONOS CLOUD Docker Machine Driver Robot"}
 committer_email=${COMMITTER_EMAIL:-"sdk@cloud.ionos.com"}
 
 function usage() {